+++ sdk/kubernetesconfiguration/azure-resourcemanager-kubernetesconfiguration/src/main/java/com/azure/resourcemanager/kubernetesconfiguration/implementation/FluxConfigOperationStatusClientImpl.java
@@ -118,11 +118,12 @@ public final class FluxConfigOperationStatusClientImpl implements FluxConfigOper
         if (operationId == null) {
             return Mono.error(new IllegalArgumentException("Parameter operationId is required and cannot be null."));
         }
+        final String apiVersion = "2024-04-01-preview";
         final String accept = "application/json";
         return FluxUtil
             .withContext(context -> service.get(this.client.getEndpoint(), this.client.getSubscriptionId(),
-                resourceGroupName, clusterRp, clusterResourceName, clusterName, fluxConfigurationName,
-                this.client.getApiVersion(), operationId, accept, context))
+                resourceGroupName, clusterRp, clusterResourceName, clusterName, fluxConfigurationName, apiVersion,
+                operationId, accept, context))
             .contextWrite(context -> context.putAll(FluxUtil.toReactorContext(this.client.getContext()).readOnly()));
     }
 
@@ -176,11 +177,11 @@ public final class FluxConfigOperationStatusClientImpl implements FluxConfigOper
         if (operationId == null) {
             return Mono.error(new IllegalArgumentException("Parameter operationId is required and cannot be null."));
         }
+        final String apiVersion = "2024-04-01-preview";
         final String accept = "application/json";
         context = this.client.mergeContext(context);
         return service.get(this.client.getEndpoint(), this.client.getSubscriptionId(), resourceGroupName, clusterRp,
-            clusterResourceName, clusterName, fluxConfigurationName, this.client.getApiVersion(), operationId, accept,
-            context);
+            clusterResourceName, clusterName, fluxConfigurationName, apiVersion, operationId, accept, context);
     }
 
     /**